refactor: move defineTableCdc to its own file
