+++ src/codegen/cloudformation/cdc/cdc.ts
@@ -2,44 +2,12 @@ import path from 'path';
 
 import {kebabCase} from 'lodash';
 
-import type {Model, Table} from '../../parser';
+import type {Model} from '../../parser';
 import type {CloudformationPluginConfig} from '../config';
 import {combineFragments} from '../fragments/combine-fragments';
-import {makeTableDispatcher} from '../fragments/table-dispatcher';
 import type {CloudFormationFragment} from '../types';
 
 import {makeHandler} from './lambdas';
-
-/** Generates CDC config for a table */
-export function defineTableCdc(
-  table: Table,
-  config: CloudformationPluginConfig,
-  {outputFile}: {outputFile: string}
-): CloudFormationFragment {
-  if (!table.hasCdc) {
-    return {};
-  }
-
-  const {dependenciesModuleId, libImportPath, tableName} = table;
-
-  const dispatcherFileName = `dispatcher-${kebabCase(tableName)}`;
-  const dispatcherFunctionName = `${tableName}CDCDispatcher`;
-  const dispatcherOutputPath = path.join(
-    path.dirname(outputFile),
-    dispatcherFileName
-  );
-
-  return combineFragments(
-    makeTableDispatcher({
-      codeUri: dispatcherFileName,
-      dependenciesModuleId,
-      functionName: dispatcherFunctionName,
-      libImportPath,
-      outputPath: dispatcherOutputPath,
-      tableName,
-    })
-  );
-}
 
 /** Generates CDC config for a model */
 export function defineModelCdc(
