+++ src/codegen/cloudformation/cdc/index.ts
@@ -1,1 +1,2 @@
 export * from './cdc';
+export {defineTableCdc} from './table';
